+++ recipe/emailpassword/api/signin.go
@@ -43,7 +43,7 @@ func SignInAPI(apiImplementation epmodels.APIInterface, options epmodels.APIOpti
 		return err
 	}
 
-	result, err := (*apiImplementation.SignInPOST)(formFields, options, &map[string]interface{}{})
+	result, err := (*apiImplementation.SignInPOST)(formFields, options, supertokens.MakeDefaultUserContextFromAPI(options.Req))
 	if err != nil {
 		return err
 	}